Atualizacao do tape3

+++ FAS/src/decorator/tape3/Bebida.java
@@ -1,20 +1,15 @@
 package decorator.tape3;
 
-
-
 public abstract class Bebida {
 
 	private String descricao;
 
-	
-	public abstract String getDescricao() ;
+	public abstract String getDescricao();
 
-	
 	public void setDescricao(String descricao) {
 		this.descricao = descricao;
 	}
 
-	
-		public abstract float preco();
-	
-	}
+	public abstract double preco();
+
+}
