+++ FAS/src/decorator/tape3/Chocolate.java
@@ -1,25 +1,23 @@
 package decorator.tape3;
 
-
-
 public class Chocolate extends CondimentosDecorator {
 
 	Bebida bebidaWrapped;
+
 	@Override
-	public String getDescricao(){
+	public String getDescricao() {
 		return this.bebidaWrapped.getDescricao() + ", Chocolate";
 	}
-		
-		/**
+
+	/**
 		 */
-		public Chocolate(Bebida bebida){
-			this.bebidaWrapped= bebida;
-		}
+	public Chocolate(Bebida bebida) {
+		this.bebidaWrapped = bebida;
+	}
 
-		@Override
-		public float preco() {
-			return 0.60f + bebidaWrapped.preco();
-		}
-
+	@Override
+	public double preco() {
+		return 0.60 + bebidaWrapped.preco();
+	}
 
 }
