+++ FAS/src/decorator/tape3/CondimentosDecorator.java
@@ -1,13 +1,11 @@
 package decorator.tape3;
 
-
-
 public abstract class CondimentosDecorator extends Bebida {
-	
+
 	@Override
 	public abstract String getDescricao();
-	
+
 	@Override
-	public abstract float preco() ;
+	public abstract double preco();
 
 }
